+++ src/Stack/Options/GhciParser.hs
@@ -52,6 +52,6 @@ ghciOptsParser = GhciOpts
              <*> switch (long "load-local-deps" <> help "Load all local dependencies of your targets")
              -- TODO: deprecate this? probably useless.
              <*> switch (long "skip-intermediate-deps" <> help "Skip loading intermediate target dependencies" <> internal)
-             <*> boolFlags True "package-hiding" "package hiding" idm
+             <*> optional (boolFlags True "package-hiding" "package hiding" idm)
              <*> switch (long "no-build" <> help "Don't build before launching GHCi" <> internal)
              <*> switch (long "only-main" <> help "Only load and import the main module.  If no main module, no modules will be loaded.")